Moved receive loop to separate thread.
Changes:
Client and Server both got their message receive loops moved to a 2nd
thread.

Replaced std::queue with moodycamel::ReaderWriterQueue to solve thread
safety issues in the Network.

Client got exitRequested moved to the game with an atomic var. It should
be in control of the lifespan.

Lots of Network changes to shore up thread safety and make the thread
work.

NetworkInputSystem now gets the number of messages waiting at the start
of execution and only processes that many to avoid running forever.

Server's network is now in charge of a ClientSet (new class, wraps the
set). Its receive thread can wait on polling the set instead of looping over
all receivers and wasting cpu.