Improved how the loop is timed.
Changes:
Changed TitleCase functions in World to camelCase.

Switched all delta timers to an accumulator system.

Switched game loop to a while so it can catch up if necessary.

Render loop is still an if, but it gives debug output if it detects
that the render was delayed.

Modified timer class to facilitate timing the execution of loops.

Client now delays for 1ms if it detects that it has enough time.