Added header, replaced deferred disconnect.
Changes:
The server's deferred disconnect has been replaced with an
immediate one.

Client class was added to the server networking scheme to take
some of the load off the Network design-wise. It's basically an
adapter for the Peer that also handles queueing outgoing messages.

Bug was fixed where SDLNet_Quit was being called before the thread
was killed, causing segfaults on exit.

Header and message batch are now being sent and received.

Added an enum for send/receive results instead of relying on bools
and nullptr for feedback.