Client now freezes or fast-forwards on adjustment
Changes:
The client used to just have an arbitrary offset that it used to
translate its tick into the server context. Now, it's using
that server context as its own tick. When it's told to adjust
backwards, it freezes for as long as needed. For forwards, it
processes multiple ticks to catch up.

Moved Game's input event stuff out to a function for an easier-to-read
loop.

Client no longer has futureOffset, or any of the history keeping
for it.

There's no more tickOffsetAdjustment, just tick and tickAdjustment.

Acceptable tickDiffs are now +1, +2, +3, anything else gets adjusted.