Separated timers, changed render timing, cleanup.
Changes:
Separated the timers on the objects that have ticks, so they can
better measure without having to account for the execution time of
eachother.

Changed the render alpha to account for the game's execution time.
This seems to be negligible most of the time, but will probably
be good going forward.

Cleaned up some comments and temporary code.