Fixed memory leak by making RAII socket classes.
Changes:
The smart pointers to SDLNet_SocketSet and TCPsocket were
leaking memory due to not calling the proper deallocation functions.
I replaced them with RAII objects.

I went on to encapsulate most of the SDLNet functionality within these
classes, so we now have a decent path towards swapping out SDLNet if we
ever care to.

The Peer/Acceptor layers are now looking like bloat that can be removed.
The Client/Server layer along with the SocketSet/TcpSocket objects
should be able to handle the responsibilities.