(Safety) Refactored how peer replication works.
Changes:
Swapped peer replication from being circular buffer-based to
queue-based and was able to make the code a lot easier to understand
because of it.

Added a server-side check for if a client has received the full world
state yet (such as when they first connect). Not working yet, though.