Fixed timeout crash.
Changes:
Fixed a crash that occurred when running against the server
because the timeout was too low.
Only the 2nd connected client would hit it because of another
bug where I was only checking timeouts if there was socket
activity.

I changed the server's receive thread to wait for 10ms if
it didn't find any activity. It's better than before, since
we used to wait on socket activity while owning the client lock,
but still need more thought.