Started dynamic npc offset, updated libSDL2pp.
Changes:
Started implementing a dynamic offset for the NPC tick. It responds to
adjustments that we get from the server.

Moved libSDL2pp back to the original repo to solve a warning on mingw.
We forked to add the libs/includes to the target, but it's
not a big deal since SDL2 also forces us to use cmake vars.