Fixed MessageSorter threading bug.
Changes:
Fixed a bug where startReceive() would fail if try_lock() either
returned false with a spurious failure, or if we tried to startReceive()
while a push was occurring. Amazingly, even when testing 100 clients,
we only ran into this situation twice. This shows that there's very
little contention for this lock.