Started trying client-side drop detection.
Changes:
Moved player data in the world into a PlayerData object.

Started working on logic for client-side player input message
drop detection.

Results:
I'm basically having to implement reliable UDP logic on top of
TCP because of higher-layer drops, it's just too complicated
and would be too hard to maintain. The server knows when it drops
a message, it's likely simpler to just send that data to the client.