(Safety) Removed drop message, cleanup.
Changes:
Removed the idea of sending a message when the server drops
a player input. It would be too complicated client-side to
rewind player position, since we only track inputs.
Instead, the server needs to send the player entity state.

Added a PlayerData class to the client, to organize the World a bit
better.