Added AoI system.
Changes:
Added an area of interest system, so only entities within a certain
range of the client will be replicated.

Renamed ClientState -> ClientSimData for clarity.

Added AreaOfInterest struct. Currently AoI is handled in 2d because I
didn't want to do extra work to get 3d when it's likely to change with
the collision system that will be eventually added.