Added comment about server exit behavior.
Changes:
Added a comment that explains why ctrl+c doesn't work.

SDL2 sends an event on quit, but our receiveCliInput thread would
have to have a non-blocking getline() for it to shut down.

Alternatively, we could just get rid of the input thread, but
it'll be useful later on for cli commands.

The real answer is probably to find a non-blocking getline() and start
handling the SQL_QUIT event from SIGINT.