Added ordered event handling.
Changes:
Added EventHandler interface for the event handling Client constructs.

Events are now passed through an ordered list of handlers, who can
choose to mark them as handled or not.

Events are polled during downtime between the PeriodicCallers.

An event filter was set up, but currently nothing needs to be filtered.