Namespaces, filesystem, sheet adding.
Changes:
Added the SpriteEditor namespace to a lot of files that were missing it.

Replaced all occurences of std::filesystem with strings. fs::path uses
wchar on Windows, which was becoming a pain since everything else uses
utf-8 and normal char.

Started working on the sprite sheet adding functionality in the UI and
model.