Refactored asset management, minor fixes.
Changes:
Fixed SDL includes to use canonical paths.

Replaced lots of SDL2pp stuff with raw SDL2 stuff. This is due to needing
to use SDL_Texture in the asset cache, which makes the raw stuff easier to use.

ResourceManager -> AssetCache

We now manage all textures and pass SDL_Textures to the UI. This lets us re-use
sprites in the UI and world without having to load them twice.

Changed how asset paths work. The asset load function doesn't know anything
about the path, and we use Paths:: to easily get the right one. This lets us
easily load both paths from the Asset directories, and user-supplied paths.

The asset cache is now injected into all UI objects that need it. This was
decided, after a lot of thinking, to be preferable to singletons or other
static access.