(Safety) Most of the rendering works again.
Changes:
Got the world and player entity rendering again.

Took another look at how entity rendering works. We were leaning
on hardcoded positions before that hid some incorrect positioning.
We now use the Y offset from the sprite, and we assume the sprite
should be horizontally centered. If this ever becomes not the case,
we'll have to add an X offset field to the editor.

Refactored tile culling, it should be easier to understand and more
correct now.

UserInterface's ghost tile rendering isn't working, but the click
behavior is.

Ran the formatter.