SpriteEditor bugfixes.
Changes:
When a sheet is removed, the stage is now made invisible.

AddSheetDialog Y offset is now cleared appropriately.

Fixed a segfault when accessing a sprite's parent sheet ref when
there's more than 1 sheet. This was caused by the sheet ref being
invalidated when the 2nd sheet is pushed into the vector.
The fix was to replace the sheet ref with the relPath, which is the
only part of the parent sheet we use.