Changed from hand-serializing to bitsery for map.
Changes:
Changed hand-serialization logic to struct/bitsery-based
serialization. This'll make it way easier to maintain, and
gives us a better path towards chunk serialization for chunk
streaming.

Added Snapshot structs to serialize into. We're doing it this way
because it gives us a single place to do the serialization, since
it's likely that we'll have divergent TileMap classes between the
server and client.

Moved serialize/deserialize to new headers.

Added measureSize() to Serialize.h.

Moved BinaryBuffer to a separate header.