Bugfixes, updated entt, comment fixes.
Known issues:
If a client sees another client disconnect and reconnect, it'll crash.
This is because we aren't currently sending disconnect messages, so the
old enitty is still there and the new client will fail to create with the
hinted ID.

Changes:
Updated entt, fixed any broken functions to use new ones.

Removed all cases of stripping the version off an entity ID before
comparing. I can't figure out why I did that but the registry seems
to take the versioned IDs as hints just fine so we should be able
to use the full thing.

Moved some things around in PlayerMovementSystem and added some comments
for clarity. Also replaced a latestReceivedTick reference with
receivedTick because it seemed to make more sense.