(Safety) Map streaming works, messages too large.
Changes:
Added ChunkUpdateSystem to the client. It seems to work, but the
messages are too large to test the "moving in range of new chunks
case.

Removed a JSON cmake flag that apparently doesn't match any flags in
that lib. Maybe it was removed?

Added map size info to ConnectionResponse so the client can initialize
the map.

Changed "Id" -> "ID" in a lot of places for consistency.

Renamed ChunkIndex to ChunkPosition to make more sense.