Added compression on server -> client messages.
Changes:
Implemented compression for Server -> Client batches over a given size.
Use SharedConfig::BATCH_COMPRESSION_THRESHOLD and
SharedConfig::BATCH_COMPRESSION_LEVEL to tune it.

Added a MAX_BATCH_SIZE to use for holding uncompressed batches.
See the comment for more info.

Added an assert to Serialize::toBuffer() since we no longer check it in
Network::send().

Changed "message" to "buffer" or "wire" in many cases in Peer, since we're not
only sending individual messages.

Replaced MessageCount in the server header with BatchSize, which reserves
the high bit for a BatchIsCompressed flag.