Render tile sort optimization, renaming.
Changes:
Added an "Area Of Interest radius" concept. It'll be useful for
determining when something is close enough to the player to care
about. We may even split it into multiple ones.

WorldSpritePreparer now only considers the tiles within the player's
radius, instead of the whole map.

Renamed Range classes to Extent.

Replaced TileMap's lengths and tile count with extents.