View calcs to world units, fixed tile flicker.
Changes:
Replaced AOI_RADIUS_TILES with VIEW_RADIUS, which uses world
units instead of tiles. The old way was simpler in this case,
but it seems useful to move towards using world units for
radiuses since we'll eventually have round things like spells.

Fixed a float precision issue that was causing tiles to flicker while
moving or zooming.

Added a separate AOI_RADIUS that is currently unused, but will
replace the current AOI setup.

Fixed an issue with screenToWorld() where the camera wasn't properly
considered. We handle the camera outside the Transforms functions
for worldToScreen(), but it seems more convenient to do it inside
for screenToWorld().