(Safety) More AOI work, added unit tests.
Changes:
Updated some debug output to be more useful.

Refactored ClientUpdateSystem to use the client's entitiesInAOI
instead of getting dirty entities. It may or may not be faster,
but it simplifies things by removing some math and leaning on
already-done math.

Changed MovementSystem so that we only update the bounding box
and set the new location in the locator if they actually moved.

Reordered the client AOI system's position in the tick so it
has that tick's movement data to process.

Fixed some issues in EntityLocator.

Fixed some issues in BoundingBox and added more comments.

Added unit tests for BoundingBox and EntityLocator. All are
passing.