Minor changes, updated bitsery.
Known issues:
If we get more than 8 entities in a single update, we crash with
a mismatch between the size that bitsery's measure function is
reporting and the actual size.

Changes:
Updated bitsery to latest to check if it fixed the problem.

Changed the server's random spawn point distribution to only
put entities in-bounds.

Changed Input's type enum to not include None in the type count
and rearranged things to match convention.