Fixed more AOI bugs.
Changes:
Added entitiesThatEnteredAOI to ClientSimData.
We needed a way to tell ClientUpdateSystem that a client needs
an update for an entity that just entered its AOI (even if
that entity doesn't have dirty inputs). Moving the local Entered
vector to be a client member seemed like the best way to do that.

Added a !PositionHasChanged check to some AOI system logic.
We had an issue where AOI lists were getting erroneous entries
because we were adding to them on the same tick that they were
getting refreshed (because their entity's position has changed).
Now, we check to make sure the entity isn't already going to
update its lists (!PositionHasChanged) before we try to maintain
them piecemeal.

ClientUpdateSystem now sends updates for any entities that just
entered a client's AOI. Also did a refactor for maintainability.