Tcp receive now supports multi-packet messages.
Changes:
Peer::receiveBytesWait() now loops until it receives all of
the requested bytes. We'll probably want to change how this is
all set up later since a misbehaved system might cause
us to hang while waiting for data, but it's fine for now.

Lots of minor comment and convention fixes.