Fixed issues with box/sprite positioning.
Changes:
For some reason, all of the sprites in SpriteData has a 384 yOffset
when it's actually 374. Perhaps I added 10 because it matches the
tile edge height and it fixed things and I was sleep deprived.
It's fixed now, and tiles/entities have their bounding boxes and
sprites lining up properly.

SpriteLayer::fixedBounds -> worldBounds to match the other structs.

Removed SharedConfig::TILE_SCREEN_EDGE_HEIGHT since it isn't needed.
It was previously being used erroneously in transforms.

Fixed a json not parsing because of moving from the assignment operator
to braces (not sure why it doesn't parse when you change it).

Fixed modelToWorldCentered not centering properly.

Updated BoundingBoxGizmo to let the gizmo's max bounds go to the edges of
the sprite (previously stopped at TILE_WORLD_WIDTH).

Updated PropertiesPanel to match the gizmo changes.

Minor fixes to match convention.