Updated AUI, removed TODO.
Changes:
Updated AUI to get tick events.

Removed a TODO about changing cursorPosition in events to be
relative to the widget's origin. Since widgets are comparing
cursorPosition to their renderExtent, and renderExtent is relative to
the screen origin, it's more straightforward to just keep using
a screen-relative cursorPosition. It also gives better debug info.
If we decide to change this for some reason in the future, we
can switch cursorPosition to be widget-relative, and use scaledExtent
instead of renderExtent within the widget event handlers.