Fixed sprite editor property setting behavior.
Changes:
Fixed an issue where the properties panel was incorrectly displaying
values when the field was set to an upper or lower bound and the
user attempted to set it to an out-of-bounds value.
The value was getting clamped and thus not changing, and the
signalling logic was such that the signal wouldn't get sent out if
the value didn't change. Now, signals are always sent out after
attempting to set values.

Fixed an issue where sprite names weren't updating properly when
the user attempted to set a non-unique name.