(Safety) Collision is partially implemented.
Changes:
Added collision to the client entity. This'll cause big
desync currently since the server doesn't yet have collision.

Fixed the windows packaging script to match the linux one.

Moves the Log::enableFileLogging() calls to the top of main, since
they were failing to log early crashes.

Changed the MovementHelpers functions to be return-based instead of
out argument-based. This is mostly for consistency with the rest
of our patterns.

Replaced some conditional checks with AM_ASSERT. They should all
hopefully be mature enough that we don't need to test them in release.

Added an out of bounds check to getTile().

BoundingBox::getCenterPosition() -> get3dCenter(). This makes it more
clear what the difference is between it and the new asEntityPosition().

Added MovementHelpers::resolveCollisions(). It currently is a super
simple ghetto collision resolver, but it'll eventually have the
full sliding collision logic.