Collision is working.
Known issues:
One time, I got magnetized to a wall. It seemed to be client-side only,
replication was looking fine on a peer. Keep an eye out.

Changes:
Moved the client's entt group pre-initialization to an EnttGroups
class to match the server.

NPCs now replicate collision properly.

Removed asEntityPosition() and moveBoundingBox(). I re-realized that
you need to consider the sprite's stage size when going between
bounding boxes and positions.

Added boundsToEntityPostion(). It isn't currently used, but it's the
correct way to get a position from a bounding box.

Comment updates.