Build and script updates.
Changes:
InstallDependencies.bat now copies the installed dependencies into
a "bin" folder and adds the folder to the system PATH.

Added a CMake exe check to PackageEngine.bat.

Moved GCC static linkage options behind a generator expression to
check for if we're compiling on GCC.

Fixed a W3 that I missed. Everything should be /W3 now.