Refactored Windows packaging logic.
Changes:
Moved from an auto-detect-based packaging strategy to just
maintaining a simple list. The auto-detect was failing on MSVC builds
and we have so few dependencies (and they're all precisely what SDL
packages so they're easy to find) that it makes more sense to just
maintain the list.

Added a Readme.txt to our Windows package that mentions where to get
the MSVC redistributable.