Fixed replication bug, profiling updates.
Changes:
Fixed a bug in player replication on the client where, when a
message was dropped, the player entity would freak out. We were
failing to update the player's bounding box after updating the
position.

Changed our mutexes to use the Tracy macro. I have yet to see them
show up in a trace, but maybe they'll be useful.

Removed some Tracy frame markings since they weren't useful.

Added some Tracy zones.

Added an info print to InputSystem when we drop a message, to match
the other case where we can drop messages.

Changed the yDistribution on random spawn points to be further from the
edge. Since we center on positions now, entities were able to spawn
outside the bounds.

Changed EventSorter::getCurrentQueue() to return a reference instead of
a pointer. Not sure why it was a pointer previously, there's no way for
the queue to be nullptr.

Changed an AM_ASSERT when an entity goes out of the world bounds to a
condition with a LOG_FATAL. This can occur in some cases and we should
still be keeping an eye out for it in release.

Made a few syntax changes to match code style.