Fixed AOI update bugs.
Background:
The ClientAOISystem had optimization logic for only updating clients
that moved during the current tick. This necessitated some extra
logic that was buggy because of relying on a false assumption.
The assumption was that, if an entity is in another entity's AOI, the
reverse is also true. This is incorrect because entity positions are
centered on their "stage", not on their bounding box. Because of this,
an entity's bounding box can be in range of another entity without
both entities being in AOI range of eachother.
We want to maintain this definition of "entity position", because we
eventually may have multiple bounding boxes on each entity. In such
a case, basing the position on the entity's "stage" makes the most
sense.

Changes:
Removed optimization logic from ClientAOISystem, now every client's AOI
is updated every tick. We know that this is performant because we
already have been load testing with clients that are constantly moving
(thus bypassing the optimization logic).

Removed PositionHasChanged since it's no longer being used.

Updated comments.