Fixed map chunk streaming bug.
Changes:
Fixed some incorrect logic in DiscreteExtent::containsPosition().
This was causing the client to not request chunks when it needed them.

Added some comments to point out where our "origin == (0, 0)"
assumptions are being made, so they're easier to find if we ever start
centering the map on (0, 0) and supporting negative origins.