Fixed a rare fatal when disconnecting.
Changes:
Fixed a LOG_FATAL that occurred when an entity disconnected while
in range of another entity, but without the other entity being in
range of the first.
We were doing extra logic to detect that we needed to send an
EntityDelete, but since we have the locator now, we can just lean
on the AOI system to figure it out and send the message.