Fixed crash at edge of map.
Changes:
Fixed a crash caused by bad math in containsExtent(). The bottomRight
position was being set to 1 greater than what it should be, causing
entities to appear to be outside of the map when they were close to
the edge.
The real math is (x + length - 1), since length is inclusive of x.

Added a minZ test to collision. Entities shouldn't be able to go through
the floor anymore.

Added a unit test for EntityLocator, which uses containsExtent().
This covers the above crash.

Started trying out using abbreviations to denote the unit when
printing map sizes and such. 'ch' == chunk, 'ce' == cell, 'u' == unit.

Minor code cleanup and changes to match code style.