Fixed crash.
Changes:
Fixed a crash caused by simulation being given a reference to the
UI's dispatcher before the UI was constructed.
We previously had the Application own the UI's dispatcher to avoid
this, but since the engine's UI is just a passthrough now, it
doesn't have a dependency on the sim so we can just construct it
first.