Minor improvements from branch.
Since we aren't merging the multiple bounding boxes branch, this
commit pulls over the improvements that were made while working on
it.

Changes:
Lots of style improvements.

Some comment improvements.

Switch our SpriteData.json parsing logic from error code to exceptions
in SpriteDataModel. This provides better debug prints and matches
the SpriteData classes.

Fixed the SpriteData.json creation path in the sprite editor. It was
previously reporting that it failed to create directories when it
was working fine.