Tile map updates now replicate automatically.
Motivation:
Previously, Server::TileUpdateSystem would only send updates in
response to update requests that it received. Now, Server::TileMap
keeps track of tiles with dirty state and TileUpdateSystem will
iterate those tiles and send updates to all nearby entities.

Changes:
TileMapBase now has a trackDirtyState parameter. Not the most
sophisticated solution, but it works.

Server::TileUpdateSystem now has a separate sendTileUpdates() function.
The tile sending logic was moved into this function, and it runs after
the "receive and apply update requests" logic and the project's
extension function. This lets all the map updates be made first, then
we can automatically send them.
Having the two functions on one class is against the usual pattern, but
it works for now and will be easy to separate if we need to later.

The TileUpdate message used to just contain one layer, but now it
contains all the layers for that tile. This is necessary in order to
know if upper layers in a tile were cleared. For example, if layers 2
and 3 were cleared, the update system would only see an empty layer 2
(cause we shrink the vectors), and the client would never get a
"clear layer 3" message.

If trackDirtyState is true, TileMapBase will push any updated tiles into
an unordered_set.

Make some of DiscretePosition's functions const for usage with STL.

Added a std::hash() specialization for TilePosition so it can be used
with std::unordered_set. I tried putting the specialization on
DiscretePosition<T> instead, but apparently that's not possible.

Added HashTools with a nice helper function from boost.