DiscreteExtent comment update.
Changes:
Updated a comment in DiscreteExtent to explain that we're trying
to avoid confusion between BoundingBox and DiscreteExtent's
concepts of "max X". BoundingBox has a member maxX, so a developer
who is used to working with it may type DiscreteExtent.maxX.
DiscreteExtent however must provide the "max X" concept as a
member function (since it's derived from x + xLength), so .maxX
would fail to compile. By instead naming it xMax(), we avoid this
confusing situation where a developer expects it to be a member
variable when it's actually a member function.