Replaced BoundingBox component with Collision.
Motivation:
We were previously set up to use the current sprite's bounding box
for collision calcs, but that approach would cause an issue if,
after calculating collision during the movement phase, we changed
the sprite. We would then have a period where the entity's
collision box has changed, but its collision hasn't been resolved.
This could cause issues in later systems.

To avoid this (and to perhaps use a more expected approach), we're
switching to an approach where character entities have a Collision
component that gets initialized, and doesn't get changed (unless
the project changes it). Thus, character entities have a consistent
collision box that doesn't constantly change based on sprite or
animation.

Changes:
Replaced character entity's BoundingBox component with a new
Collision component that holds both the modelBounds and worldBounds.

Updated a bunch of systems and EntityLocator to use Collision.

Added a comment explaining that we're now using the default character
sprite's bounding box as the "standard collision" for all character
entities.