Updated dependency script, readme.
Changes:
Updated the readme to mention Repose.

Updated the dependency install script to set environment vars
so we don't need the user to set the paths in CMakeSettings.txt.
This is a pretty easy setup, but we can improve this further later
by removing the install script and just handling it in cmake.