Fixed invisible player bug.
Changes:
Updated entt to to be sure we had the latest bugfixes.

Changed an unsigned int to Uint32 to match the used type.

Fixed entt group call that had an exclude as the first parameter.
If you only want to exclude, you need to include empty braces as
the first parameter.

Fixed a unit test include path.