Massive tile map/sprite data improvement.
Motivation:
Previously, each tile in the tile map just held layers of sprites.
This was fine for a while, but it was recognized that, by bringing
the tile map to a higher level of abstraction, we could make the
experience of working with it a lot more friendly.
This "higher level of abstraction" is that of floors, walls, etc.
Actual concepts of what goes on a tile map, instead of generic
sprites. The goal with this change is to let developers think at
the level of "I want to add a wall to this tile", instead of "I
need to add a wall sprite to layer 3 and update any surrounding
sprites".

TODO:
Rendering is still partially commented out.

Changes:
Refactored SpriteDataBase/SpriteData classes to use the new
higher-level sprite concepts.

Replaced TileUpdate/TileUpdateRequest messages with more specific
messages: TileAddLayer, TileRemoveLayer, TileClearLayers,
TileExtentClearLayers. These were necessary because we no longer
are working with a homogenous array of sprites, but they provide
the added benefit of opening a path for us to send large operations
as a single message. TileExtentClearLayers is the current example
of that, and we can add more in the future.

Refactored the TileUpdateSystems, ChunkUpdateSystem, and
ChunkStreamingSystem to match the new tile map and use the new
messages.

Made some enum changes for consistency. When relevant, each enum
should follow the patterns: Be a Uint8 (for in-memory size in
the tile map), be a struct-wrapped normal enum (to avoid constantly
casting an enum class. not necessary if never used as index or size),
have a Count value (instead of e.g. NumTypes), don't specify numbers
unless necessary (so Count stays correct if values are added).

Gave SpriteData a dedicated directory.

SpriteDataBase::get() -> getSprite().

Big TileMapBase/TileMap refactors.

Changed ISimulationExtension:isTileUpdateValid() to isExtentEditable().
It's more restricted so we may need to change it in the future, but
we have multiple types of update messages now so the old one wasn't
workable. Maybe in the future change we'll pass the updates as a
variant?

Updated ChunkSnapshot/ChunkWireSnapshot/TileSnapshot to work with the
tile map changes.

Changed Rotation::Direction to be an unsigned enum so it fit with
Wall::Type in our generic message. MovementHelpers is the only place
that needed the special values for the input math, so it now has
a translation function.

Refactored SpriteData, updated parsing logic.

Big Tile refactor to handle the new concepts.

Fixed some things like ChunkExtent that were a class when they should
have been a struct.