Fixed wall bugs, add World dep to UI.
Changes:
Gave the UserInterface access to the sim's World object through
UserInterfaceExDependencies. worldSignals is still going to be
used for when the sim has an "event" to push to the UI, but the
UI now also has the option of observing data from the sim state.

Fixed some bugs with how wall tile layers were handled.