Implemented all sprite set views, bugfixes.
Changes:
Added a fix for the "find a unique name" loop when adding sprite sets.
I had added it for floors, but forgot to do the rest.

Turned the Floor edit stage/properties window into generic classes to
handle any of the sprite sets. This means we don't need to duplicate
and maintain all of that code for each set type.

Fixed a bug where LibraryAddMenu was behind the stage.