Wall tile improvements, minor changes.
Changes:
Minor changes for style consistency.

When modifying walls that form a down-facing corner, if both walls
are changed to no longer match the NW fill, it'll be changed to
match the newer wall.

Fixed an issue where adding a West wall to a tile with a North wall
resulted in a NE fill that matched the new wall. Now it matches the
old wall.

Improved SpriteEditor Library list item selection.