Fixed entity handling bugs.
Changes:
Updated entt to get latest bugfixes.

Entt now (partially?) registers entities with an ID less than
the hint you pass to create(). This created a crash during NPC
creation when we check if the ID is available. Instead, we now
just try to create the entity and check afterwards if it got the
right ID.

Fixed a bug where, after adding a dynamic object and rejoining,
the new client entity would get the object's collision and warp
around. This was caused by using invalidated component references.
Each of the component creation areas now features a big note now to
be careful with holding onto references.