(Safety) Started implementing Interactions.
Changes:
Added Interactions components to client entities and dynamic objects.

Replaced World dependency in SimulationExtension with Simulation.
You can get the World from the Simulation, and this lets us expose
extra sim functionality like interaction events.

Replaced Ray's origin members with a Position struct. The idea is,
if we already have a struct that encompasses the concept, to use it.
We don't have one for direction vectors, so that stays as separate fields
for now.

Added Cylinder struct and started using it in appropriate existing places.

Fixed sim extension passing issue. We were previously passing it as a
parameter in some spots, when it isn't actually available at construction
time. Now we appropriately grab it through a setExtension() call where needed.

Moved DynamicObject construction logic into a World function, since we
need to construct dynamic objects from multiple places.

Added data flow for processing/dispatching/receiving interactions.

Added a listener for entity destruction in the server, so we can clean them
out of the entity locator without accidentally forgetting to.

Refactored the entity locator to have 2 types of function: one for
getting entities based on position, and another for getting them
based on collision box. This lets us fix a bug where an entity would be in
range of another without the opposite being true, causing us to send sprite
updates to a client for an entity that never got init'd.