Automatic component updates are working.
Changes:
Added a ComponentUpdate message and systems. When a component
in the ObservedComponentMap is updated, we now send the
associated components automatically, and replicate them
automatically on the other side.

Replaced EntityType with IsClientEntity. We want to think of
entities as a generic thing now, with interactions and
behaviors defining their type.

Removed SpriteUpdateSystem and AnimationStateNeedsSync (handled by
automatic updates now).

Updated the ISimulationExtensions to go better with the changes.

Replaced client->server component update requests with a generic
ComponentUpdateRequest message.

Added support for empty component types to ClientAOISystem.

Renamed ComponentSyncSystem -> ComponentUpdateSystem.

Renamed DynamicObjectInitRequest -> EntityInitRequest.

Refactored the ISimulationExtension validation functions and
added new ones.

In messages where we had comments for "Replicated data", renamed
it to "Networked data" to be less confusing since we now have an
unrelated ReplicatedComponents type.

Added RemoveCvRef, a meta helper.