Moved movement back to manual updates.
Changes:
After benchmarking, data increased by 50% and speed decreased by 50%
for movement updates after moving to auto-updates. By putting
movement back on manual, we gain the performance back while
maintaining 95% of the convenience that the auto-updates bring.

Changed to a new manual update setup for component change requests.
We still want to handle what we can in the engine since we'll likely
be able to support everything needed for build mode, but the full
variant setup seemed unnecessary.

Changed ObservedComponentMap to a flat list. The one -> many
relationship seemed unnecessary with movement gone, and this simplifies
things.