Added AI, component refactor.
Changes:
UserErrorString -> SystemMessage so we can use it for more things.

Cleaned up the onAnimationStateUpdated functions.

Added AILogic class, AIBehavior component, and AISystem for processing.

Added a place to add lua bindings.

Moved some common component creation logic into a few World functions.
In the process, removed the weird variant vector interface.

Added Simulation::getLua() so the project can access it.

Replaced EntityInitRequest's component variant vector with a few
hardcoded components. It seems like they're all that'll be needed,
and the variant interface was weird to work with.

Small change to EngineInteractionType to see if we like it better.
Hopefully the engine won't actually end up being useful here, so we
can just remove it and change ProjectInteractionType to an
InteractionType enum class. That'd still have to cast to Uint8
in the engine, though.